requirements.txt
asgiref==3.6.0
cffi==1.15.1
cryptography==40.0.1
Django==4.2
django-environ==0.10.0
django-filter==23.1
django-rest-knox==4.2.0
djangorestframework==3.14.0
Markdown==3.4.3
psycopg==3.1.8
psycopg-binary==3.1.8
pycparser==2.21
pytz==2023.3
sqlparse==0.4.3
typing_extensions==4.5.0
Pillow==9.5.0


